feat: move `get_study_locus_id` inside studylocus
