+++ src/otg/dataset/study_locus.py
@@ -221,6 +221,31 @@ class StudyLocus(Dataset):
         """
         return f.filter(credible_set, lambda x: x["is95CredibleSet"])
 
+    @staticmethod
+    def get_study_locus_id(study_id_col: Column, variant_id_col: Column) -> Column:
+        """Hashes a column with a variant ID and a study ID to extract a consistent studyLocusId.
+
+        Args:
+            study_id_col (Column): column name with a study ID
+            variant_id_col (Column): column name with a variant ID
+
+        Returns:
+            Column: column with a study locus ID
+
+        Examples:
+            >>> df = spark.createDataFrame([("GCST000001", "1_1000_A_C"), ("GCST000002", "1_1000_A_C")]).toDF("studyId", "variantId")
+            >>> df.withColumn("study_locus_id", StudyLocus.get_study_locus_id(*[f.col("variantId"), f.col("studyId")])).show()
+            +----------+----------+--------------------+
+            |   studyId| variantId|      study_locus_id|
+            +----------+----------+--------------------+
+            |GCST000001|1_1000_A_C| 7437284926964690765|
+            |GCST000002|1_1000_A_C|-7653912547667845377|
+            +----------+----------+--------------------+
+            <BLANKLINE>
+
+        """
+        return f.xxhash64(*[study_id_col, variant_id_col]).alias("studyLocusId")
+
     @classmethod
     def from_parquet(cls: type[StudyLocus], session: Session, path: str) -> StudyLocus:
         """Initialise StudyLocus from parquet file.
