+++ src/otg/method/window_based_clumping.py
@@ -11,7 +11,6 @@ from pyspark.ml.linalg import DenseVector, VectorUDT
 from pyspark.sql.window import Window
 
 from otg.common.spark_helpers import calculate_neglog_pvalue
-from otg.common.utils import get_study_locus_id
 from otg.dataset.study_locus import StudyLocus
 
 if TYPE_CHECKING:
@@ -336,6 +335,7 @@ class WindowBasedClumping:
             .drop("isLead", "negLogPValue", "cluster_id")
             # assign study-locus id:
             .withColumn(
-                "studyLocusId", get_study_locus_id(f.col("studyId"), f.col("variantId"))
+                "studyLocusId",
+                StudyLocus.get_study_locus_id(f.col("studyId"), f.col("variantId")),
             )
         )
